Separando os componentes e adicioando pastas
